Create separate module for authentication
