+++ frontend/src/dashboard/MainMenu.vue
@@ -1,17 +1,17 @@
 <template>
-    <div>
-        <Authentication />
-    </div>
+  <div>
+    <Authentication />
+  </div>
 </template>
 
 <script lang="ts">
-import { Component, Vue } from 'vue-property-decorator';
-import Authentication from './Authentication.vue';
+import { Component, Vue } from "vue-property-decorator";
+import Authentication from "../authentication/Authentication.vue";
 
 @Component({
-    components: {
-        Authentication
-    }
+  components: {
+    Authentication
+  }
 })
 export default class MainMenu extends Vue {}
 </script>
